+++ CHANGELOG.md
@@ -4,9 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.7.0]
 ### Added
 - parameter to configure docker stop timeout duration.
+- parameter to configure aws volume iops, by [ttousai](https://github.com/ttousai).
+- parameter to configure gcp scopes, by [imranismail](https://github.com/imranismail).
+- metrics to track server boot errors
+- metrics to track server boot time
+- metrics to track server installation errors
+- metrics to track server installation time
+- metrics to track server creation errors
+- metrics to track server creation time
+
+### Fixed
+- do not run docker stop if the instance was not created.
+- do not run docker stop if the instance was not assigned an IP.
 
 ## [1.6.1]
 ### Added